requirements.txt
streamlit
PyMuPDF
python-docx
ffmpeg-python
llama-cpp-python
requests
